requirements.txt
pyyaml>=6.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
